+++ pkg/scalers/cron_scaler.go
@@ -73,21 +73,27 @@ func parseCronMetadata(config *ScalerConfig) (*cronMetadata, error) {
 	} else {
 		return nil, fmt.Errorf("no timezone specified. %s", config.TriggerMetadata)
 	}
+	parser := cron.NewParser(cron.Minute | cron.Hour | cron.Dom | cron.Month | cron.Dow)
 	if val, ok := config.TriggerMetadata["start"]; ok && val != "" {
-		if strings.Contains(val, "-") {
-			return nil, fmt.Errorf("error parsing start schedule. %s: range or hyphenated inputs are not allowed", config.TriggerMetadata)
+		_, err := parser.Parse(val)
+		if err != nil {
+			return nil, fmt.Errorf("error parsing start schedule: %s", err)
 		}
 		meta.start = val
 	} else {
 		return nil, fmt.Errorf("no start schedule specified. %s", config.TriggerMetadata)
 	}
 	if val, ok := config.TriggerMetadata["end"]; ok && val != "" {
-		if strings.Contains(val, "-") {
-			return nil, fmt.Errorf("error parsing end schedule. %s: range or hyphenated inputs are not allowed", config.TriggerMetadata)
+		_, err := parser.Parse(val)
+		if err != nil {
+			return nil, fmt.Errorf("error parsing end schedule: %s", err)
 		}
 		meta.end = val
 	} else {
 		return nil, fmt.Errorf("no end schedule specified. %s", config.TriggerMetadata)
+	}
+	if meta.start == meta.end {
+		return nil, fmt.Errorf("error parsing schedule. %s: start and end can not have exactly same time input", config.TriggerMetadata)
 	}
 	if val, ok := config.TriggerMetadata["desiredReplicas"]; ok && val != "" {
 		metadataDesiredReplicas, err := strconv.Atoi(val)