+++ pkg/scaling/executor/scale_jobs.go
@@ -1,11 +1,25 @@
+/*
+Copyright 2021 The KEDA Authors
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
 package executor
 
 import (
 	"context"
 	"sort"
 	"strconv"
-
-	"github.com/kedacore/keda/v2/pkg/eventreason"
 
 	"github.com/go-logr/logr"
 	batchv1 "k8s.io/api/batch/v1"
@@ -14,7 +28,8 @@ import (
 	"sigs.k8s.io/controller-runtime/pkg/client"
 	"sigs.k8s.io/controller-runtime/pkg/controller/controllerutil"
 
-	kedav1alpha1 "github.com/kedacore/keda/v2/api/v1alpha1"
+	kedav1alpha1 "github.com/kedacore/keda/v2/apis/keda/v1alpha1"
+	"github.com/kedacore/keda/v2/pkg/eventreason"
 	version "github.com/kedacore/keda/v2/version"
 )
 
@@ -310,7 +325,7 @@ func (e *scaleExecutor) deleteJobsWithHistoryLimit(logger logr.Logger, jobs []ba
 		deleteOptions := &client.DeleteOptions{
 			PropagationPolicy: &deletePolicy,
 		}
-		err := e.client.Delete(context.TODO(), j.DeepCopyObject(), deleteOptions)
+		err := e.client.Delete(context.TODO(), j.DeepCopy(), deleteOptions)
 		if err != nil {
 			return err
 		}